+++ BambooTracker/io/instrument_io.hpp
@@ -33,6 +33,9 @@ private:
 	static AbstractInstrument* loadY12File(std::string path,
 										   std::weak_ptr<InstrumentsManager> instMan,
 										   int instNum);
+	static AbstractInstrument* loadINSFile(std::string path,
+										   std::weak_ptr<InstrumentsManager> instMan,
+										   int instNum);
 
 public:
 	static AbstractInstrument* loadWOPNInstrument(const WOPNInstrument &srcInst,